backend/requirements.txt
Django==6.0.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
openpyxl==3.1.5
reportlab==4.2.2
